refactor: place the extras in a dedicated folder
