+++ geetools/extra/clouds.py
