+++ tests/test_ImageCollection/serialized_test_reduce_interval.yml
@@ -602,7 +602,7 @@ values:
           valueReference: '24'
       functionName: String.cat
   '4':
-    constantValue: 8445b85e2c384324a9893cdceabdf9b1
+    constantValue: 572b1f681142457faef6ded01b1a4cfe
   '5':
     functionInvocationValue:
       arguments:
