+++ tests/test_ImageCollection/serialized_test_reduce_interval_properties.yml
@@ -602,7 +602,7 @@ values:
           valueReference: '24'
       functionName: String.cat
   '4':
-    constantValue: 46aee25b28d64238a3655d1c5d11bf60
+    constantValue: 2f5cc01048264ac189cc276a20c955a8
   '5':
     functionInvocationValue:
       arguments:
