+++ tests/test_ImageCollection/serialized_test_reduce_interval_with_multi_output_reducer.yml
@@ -630,7 +630,7 @@ values:
           valueReference: '12'
       functionName: String.cat
   '4':
-    constantValue: 83868958d4844166881fa7996fe58efa
+    constantValue: d17b46830f77402581c178123abcfed2
   '5':
     functionInvocationValue:
       arguments:
