+++ tests/test_ImageCollection/serialized_test_reduce_interval_with_reducer.yml
@@ -580,7 +580,7 @@ values:
           valueReference: '23'
       functionName: String.cat
   '4':
-    constantValue: 36a1f1d73f7842c893e1edc1c912835f
+    constantValue: 6f3c2db3d96e49a393a38f34a0accb9c
   '5':
     functionInvocationValue:
       arguments:
