+++ tests/test_ImageCollection/serialized_test_reduce_interval_without_original_names.yml
@@ -628,7 +628,7 @@ values:
           valueReference: '12'
       functionName: String.cat
   '4':
-    constantValue: 73c17dea05234a30a3c444de2ead0825
+    constantValue: b53e416094894d01b28256699afc1a21
   '5':
     functionInvocationValue:
       arguments:
