+++ tests/test_ImageCollection/serialized_test_reduce_regions_by_date_property.yml
@@ -183,7 +183,7 @@ values:
             functionName: Date.format
       functionName: Element.set
   '16':
-    constantValue: d7d057ad6ffd4d9f95d6e7a558786d9e
+    constantValue: 82ea39cf40ed446fb795c7b834d2f830
   '17':
     constantValue: system:time_start
   '18':
