+++ tests/test_ImageCollection/serialized_test_reduce_regions_by_dates.yml
@@ -183,7 +183,7 @@ values:
             functionName: Date.format
       functionName: Element.set
   '16':
-    constantValue: e1cd902c75fc48df8bc4dff36ba10e8c
+    constantValue: f5810196575b4ba5a7237b41f4479513
   '17':
     constantValue: system:time_start
   '18':
