+++ tests/test_ImageCollection/serialized_test_reduce_regions_by_doy.yml
@@ -202,7 +202,7 @@ values:
             functionName: Date.format
       functionName: Element.set
   '16':
-    constantValue: b7a80e7f9dcf476f83a06684a710c8f9
+    constantValue: cc3449bfee7f4618b5ef7fcb05c30952
   '17':
     constantValue: system:time_start
   '18':
